insignificant tweaks

+++ src/libprofiler/eventlist.cpp
@@ -415,16 +415,17 @@ void EventList::dumpEvents(char* path, char * ip_filename)
 printf("Timer start %lu \n",this->gpu_timer_start) ;
     // Write the events to file with each field separated by semicolons
     for(int i = 0; i < (int)this->event_list.size(); i++) {
-        fprintf(fp, "%s;\t%s;\t%f;\t%f;\t%f;\t%f\n",
+//        fprintf(fp, "%s;\t%s;\t%f;\t%f;\t%f;\t%f\n",
+	    fprintf(fp, "%s;\t%s;\t%lu;\t%lu;\t%lu;\t%lu\n",
             this->event_list[i]->type,
             this->event_list[i]->name, 
-            float( getEventValue(this->event_list[i]->event,
+            long( getEventValue(this->event_list[i]->event,
                 	CL_PROFILING_COMMAND_QUEUED) - this->gpu_timer_start ),
-	    float( getEventValue(this->event_list[i]->event,
+	    long( getEventValue(this->event_list[i]->event,
 	                CL_PROFILING_COMMAND_SUBMIT) - this->gpu_timer_start ),
-            float( getEventValue(this->event_list[i]->event,
+            long( getEventValue(this->event_list[i]->event,
         	        CL_PROFILING_COMMAND_START) - this->gpu_timer_start ),
-            float( getEventValue(this->event_list[i]->event,
+            long( getEventValue(this->event_list[i]->event,
         	        CL_PROFILING_COMMAND_END) - this->gpu_timer_start )  
 	);
 /*
@@ -459,7 +460,7 @@ printf("Timer start %lu \n",this->gpu_timer_start) ;
     if(event_list.size() > 0)
     {
     	long start_t = getEventValue(this->event_list[0]->event, CL_PROFILING_COMMAND_QUEUED) - this->gpu_timer_start ;
-		long end_t = getEventValue(this->event_list[event_list.size() - 1]->event,CL_PROFILING_COMMAND_END)	- this->gpu_timer_start;
+	long end_t = getEventValue(this->event_list[event_list.size() - 1]->event,CL_PROFILING_COMMAND_END)	- this->gpu_timer_start;
 		printf("Time first %lu \t last %lu \t Diff %f \n",start_t, end_t, float((end_t - start_t))/(1000.0f*1000.0f) );
     }
     else
